commit with follow and unfollow fonctions  accomplished

+++ controllers/user.controller.js
@@ -49,8 +49,13 @@ module.exports.updateUser = async (req, res) => {
 };
 
 module.exports.deleteUser = async (req, res) => {
-  if (!ObjectId.isValid(req.params.id))
-    return res.status(400).send("ID unknown " + req.params.id);
+  if (
+    !ObjectId.isValid(req.params.id) ||
+    !ObjectId.isValid(req.body.idToFollow)
+  )
+    return res
+      .status(400)
+      .send("ID unknown " + req.params.id + "or" + req.body.idToFollow);
   try {
     const post = await userModel.findByIdAndDelete(req.params.id);
 
@@ -106,12 +111,51 @@ module.exports.follow = async (req, res) => {
 };
 
 module.exports.unFollow = async (req, res) => {
-  if (!ObjectId.isValid(req.params.id))
-    return res.status(400).send("ID unknown " + req.params.id);
+  if (
+    !ObjectId.isValid(req.params.id) ||
+    !ObjectId.isValid(req.body.idToUnFollow)
+  ) {
+    return res
+      .status(400)
+      .send("ID inconnu: " + req.params.id + " ou " + req.body.idToUnFollow);
+  }
 
   try {
+    console.log("ID de l'utilisateur qui ne suit plus : ", req.params.id);
+    console.log(
+      "ID de l'utilisateur à ne plus suivre : ",
+      req.body.idToUnFollow
+    );
+
+    // Retirer de followers
+    const follower = await userModel.findByIdAndUpdate(
+      req.params.id,
+      { $pull: { followers: req.body.idToUnFollow } },
+      { new: true }
+    );
+
+    if (!follower) {
+      return res.status(404).send({ message: "Utilisateur non trouvé" });
+    }
+
+    console.log("Followers après retrait : ", follower.followers);
+
+    // Retirer de following
+    const following = await userModel.findByIdAndUpdate(
+      req.body.idToUnFollow,
+      { $pull: { following: req.params.id } },
+      { new: true }
+    );
+
+    if (!following) {
+      return res.status(404).send({ message: "Utilisateur non trouvé" });
+    }
+
+    console.log("Following après retrait : ", following.following);
+
+    res.status(200).json({ follower, following });
   } catch (err) {
-    console.log("l'erreur est " + err);
-    res.status(500).json({ message: "message error" + err });
+    console.log("L'erreur est : " + err);
+    res.status(500).json({ message: "Erreur : " + err });
   }
 };
